UIROLES-63 correctly assign/unassign users from roles
The main refactor is to move the role-manipulation handler from
`useEffect` directly into the callback from Pluggable, which is
effectively an event-handler, since React recommends not using Effects
to handler user events.

The secondary refactor is to issue per-user requests to retrieve their
assigned roles, rather than retrieving all the user-role records via the
`useUserRolesByUserIds` hook. That's unfortunate since it means issuing
a query for each user whose roles are changing, but I couldn't wrap my
head around how to leverage that hook given it needs to be called in an
event-handler.

Refs UIROLES-63, UIROLES-43