UIROLES-53 properly handle react-query 'onError' errors (#69)
When a react-query request returns non-ok, it passes the resolved
fetch as the first argument to its `onError` handler. The previous
implementation of `useErrorCallout` did not handle this, and instead
naively expected a string but did not do any input validation.

`callout.sendCallout()`, for its part, _also_ does absolutely zero input
validation and, due to its location high in the hierarchy of rendered
components, takes down all of Stripes when it fails to render. That
ain't great (hence STCOM-1300).

Refs UIROLES-53, STCOM-1300